pyproject.toml
[project]
name = "wordlebench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.4.4",
    "openai>=2.20.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.0",
]
